Use string instead of float to determine mobility
Comparing floats was a bad idea and did not work accurately. 
Instead, store the current location data to a string and run through the loop checking for a change to the previous location value.

TODO: Use real-world testing to see if gps.location.speed() reports accurate data. If so, we can use this to determine if we are mobile, in addition to, or rather than using location data.