Use gps.speed.kmph() to determine if mobile
Instead of comparing the current latitude and longitude to the previous 4 cycles, use gps.speed.kmph()
Ensure that we check if gps.speed.kmph() is over 5, as sometimes while we are stationary it can be reported that we are moving. It can sometimes report 0-4.5km/hr when in fact we are stopped. I believe this is due to gps fix or fix updates.

I have also removed while(!Serial) as this seems to stop everything - because the Arduino is not plugged into a PC when on the field!