Add LED error codes
Added ledErrorCode() function, which will flash the red and blue onboard LED to indicate an error has occurred. 
This provides a visual display of issues that have occurred on the field, where a serial port connection is not possible (i.e. not connected to a PC).

The red LED will long flash once, followed by a short flash on the blue LED to indicate the error code. 
For example:
1 long flash = 10th unit
1 short flash = 1 unit
1 long flash + 1 short flash = error code 11 (serial error)

The following errors are currently implemented:
11: Serial error (No communication from GPS unit)
12: SD Card error (No card, wrong chip select pin, or wiring error)
13: Waiting for location sync
14: Waiting on date sync
15: Waiting on time sync
21: SD card write error