Improve data points for logging and accuracy
Added speed, location age and satellite values to the log file. 

This should improve accuracy, as often when the data points are inaccurate it is a result of poor satellite connection or a high location age (>1850ms).
As a result, we also check the values of location age and satellite count before we start logging.

More testing is required, but there seems to be a pattern in bad data points that correlates to these values.